Add normalize()
